README and other updates
